% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam1.R
\name{sam1}
\alias{sam1}
\title{One Way Tabulation That Includes the Missing values (Columnswise)}
\usage{
sam1(row, decimal = 1, sort.group = c(FALSE, "decreasing", "increasing"),
  cum.percent = !any(is.na(row)), graph = FALSE, missing = TRUE,
  bar.values = c("frequency", "percent", "none"), horiz = FALSE, cex = 1,
  cex.names = 1, main = "auto", xlab = "auto", ylab = "auto",
  col = "auto", gen.ind.vars = FALSE, ...)
}
\arguments{
\item{row}{The factors to be analysed}
}
\value{
value
}
\description{
One Way Tabulation That Includes the Missing values (Columnswise)
}
