% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam2.R
\name{sam2}
\alias{sam2}
\title{Two Way Tabulation That Includes the Missing values (Columnswise)}
\usage{
sam2(row, column, decimal = 2, percent = c("col", "row"), missing = TRUE,
  sort.group = c(decreasing, "increasing"), cum.percent = !any(is.na(row)),
  las = 0, main = "auto", xlab = "auto", ylab = "auto", col = "auto",
  ...)
}
\arguments{
\item{row}{The factors to be analysed}

\item{column}{The groups to be considered}
}
\value{
value
}
\description{
Two Way Tabulation That Includes the Missing values (Columnswise)
}
