% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samhtml.R
\name{samhtml}
\alias{samhtml}
\title{Two Way Tabulation with HTML Table Output}
\usage{
samhtml(factor, group)
}
\arguments{
\item{factors}{The factors to be analysed}

\item{groups}{The groups to be considered}
}
\value{
value
}
\description{
Two Way Tabulation with HTML Table Output
}
