% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samkable.r
\name{samkable}
\alias{samkable}
\title{Two Way Tabulation with Kable Table Output}
\usage{
samkable(factor, group)
}
\arguments{
\item{factors}{The factors to be analysed}

\item{groups}{The groups to be considered}
}
\value{
value
}
\description{
Two Way Tabulation with Kable Table Output
}
