% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampct.r
\name{sampct}
\alias{sampct}
\title{Two Way Analysis That includes NA's}
\usage{
sampct(row, column, decimal = 1, percent = c("both", "col", "row"),
  graph = FALSE, las = 0, main = "auto", xlab = "auto", ylab = "auto",
  col = "auto", ...)
}
\arguments{
\item{row}{The factors to be analysed}

\item{column}{the groups to be considered}
}
\value{
value
}
\description{
Two Way Analysis That includes NA's
}
