% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\alias{table2.default}
\alias{table2.formula}
\title{Generate an HTML table of descriptive statistics.}
\usage{
table2(x, ...)

\method{table2}{default}(x, labels, groupspan = NULL, rowlabelhead = "",
  transpose = FALSE, topclass = "Rtable2", footnote = NULL,
  render = render.default, ...)

\method{table2}{formula}(x, data, overall = "Overall", rowlabelhead = "",
  transpose = FALSE, droplevels = TRUE, topclass = "Rtable2",
  footnote = NULL, render = render.default, ...)
}
\arguments{
\item{x}{An object, typically a \code{formula} or list of \code{data.frame}s.}

\item{...}{Further arguments, passed to \code{render}.}

\item{labels}{A list containing labels for variables, strata and groups (see Details).}

\item{groupspan}{A vector of integers specifying the number of strata to group together.}

\item{rowlabelhead}{A heading for the first column of the table, which contains the row labels.}

\item{transpose}{Logical. Should the table be transposed (i.e. strata as
rows and variables as columns)?}

\item{topclass}{A class attribute for the outermost (i.e. \code{<table>}) tag.}

\item{footnote}{A character string to be added as a footnote to the table.
The default \code{NULL} causes the footnote to be omitted.}

\item{render}{A function to render the table cells (see Details).
immediately displayed? Otherwise an HTML fragment is printed to
\code{\link{stdout}}.}

\item{data}{For the formula interface, a \code{data.frame} from which the
variables in \code{x} should be taken.}

\item{overall}{A label for the "Overall" column. Specify \code{NULL} or
\code{FALSE} to omit the column altogether.}

\item{droplevels}{Should empty factor levels be dropped?}
}
\value{
An object of class "table2".
}
\description{
}
