% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampct.r
\name{Chi.trend.test}
\alias{Chi.trend.test}
\title{Two Way Analysis That includes NA's}
\usage{
Chi.trend.test(binomial,ordinal,table)
}
\arguments{
\item{binomial}{The factors to be analysed}
\item{ordinal}{the groups to be considered}
\item{ordinal}{a table}
}
\value{
value
}
\description{
Two Way Analysis That includes NA's
}
